+++ src/ui/Input.h
@@ -152,16 +152,16 @@ namespace ui
 	inline bool slider_field(Widget& parent, cstring name, AutoStat<T> value, bool reverse = false) { return field([&](Widget& self) { return slider_input<T>(self, value); }, parent, name, reverse); }
 
 	export_ func_ inline bool flag_field(Widget& parent, cstring name, uint32_t& value, uint8_t shift, bool reverse = false) { return field([&](Widget& self) { return flag_input(self, value, shift); }, parent, name, reverse); }
-	export_ func_ inline bool radio_field(Widget& parent, cstring name, array<cstring> choices, size_t& value, bool reverse = false) { return field([&](Widget& self) { return radio_switch(self, choices, value); }, parent, name, reverse); }
-	export_ func_ inline bool dropdown_field(Widget& parent, cstring name, array<cstring> choices, size_t& value, bool reverse = false) { return field([&](Widget& self) { return dropdown_input(self, choices, value); }, parent, name, reverse); }
-	export_ func_ inline bool typedown_field(Widget& parent, cstring name, array<cstring> choices, size_t& value, bool reverse = false) { return field([&](Widget& self) { return typedown_input(self, choices, value); }, parent, name, reverse); }
+	export_ func_ inline bool radio_field(Widget& parent, cstring name, array<cstring> choices, uint32_t& value, bool reverse = false) { return field([&](Widget& self) { return radio_switch(self, choices, value); }, parent, name, reverse); }
+	export_ func_ inline bool dropdown_field(Widget& parent, cstring name, array<cstring> choices, uint32_t& value, bool reverse = false) { return field([&](Widget& self) { return dropdown_input(self, choices, value); }, parent, name, reverse); }
+	export_ func_ inline bool typedown_field(Widget& parent, cstring name, array<cstring> choices, uint32_t& value, bool reverse = false) { return field([&](Widget& self) { return typedown_input(self, choices, value); }, parent, name, reverse); }
 	export_ func_ inline bool color_field(Widget& parent, cstring name, Colour& value, bool reverse = false) { return field([&](Widget& self) { return color_toggle_edit(self, value); }, parent, name, reverse); }
 	export_ func_ inline void color_display_field(Widget& parent, cstring name, const Colour& value, bool reverse = false) { field([&](Widget& self) { color_display(self, value); return false; }, parent, name, reverse); }
 
 	template <class T>
 	inline bool enum_input(Widget& parent, T& value)
 	{
-		size_t index = enum_index(Ref(&value));
+		uint32_t index = enum_index(Ref(&value));
 		//ui::radio_switch(parent, meta(value).m_enum_names, index);
 		if(ui::dropdown_input(parent, to_array(enu<T>().m_names), index))
 		{
